requirements.txt
Flask==0.10.1
psycopg2==2.6.1 
Flask-SQLAlchemy===2.1
Flask-Migrate==1.8.0
simplejson==3.8.2

